+++ gamepad/src/main/java/com/serenegiant/gamepad/IGamePad.java
@@ -4,11 +4,11 @@ import static com.serenegiant.gamepad.GamePadConst.*;
 
 public abstract class IGamePad {
 	protected final Object mKeySync = new Object();
-	protected final int[] mDownTimes = new int[KEY_NUMS];
+	protected final long[] mDownTimes = new long[KEY_NUMS];
 	protected final boolean[] mIsDowns = new boolean[KEY_NUMS];
 	protected boolean mModified = true;
 
-	public int[] downTimes() {
+	public long[] downTimes() {
 		updateState(mIsDowns, mDownTimes, false);
 		return mDownTimes;
 	}
@@ -23,6 +23,6 @@ public abstract class IGamePad {
 	 * @param downs KEY_NUMS個以上確保しておくこと
 	 * @param down_Times KEY_NUMS個以上確保しておくこと
 	 */
-	public abstract void updateState(final boolean[] downs, final int[] down_Times, final boolean force);
+	public abstract void updateState(final boolean[] downs, final long[] down_Times, final boolean force);
 
 }
